[CBRD-26761] Move ftab_set from parallel_query to storage layer
Relocate src/query/parallel/px_ftab_set.hpp to src/storage/ftab_set.hpp
and drop the parallel_query namespace so the type becomes top-level ftab_set.
All five existing callers (external_sort.c, btree_load.{c,h},
px_heap_scan_input_handler_ftabs.hpp) migrate atomically - no using-alias
residue. Mechanical relocation; no behavior change.

Also declare collect_strided_vpids in the new header; its implementation
lands in the next commit (storage/ftab_set.cpp).
